Finalized all endpoints besides enroll student, added even more error handling, fixed krakend endpoints

+++ api.py
@@ -28,10 +28,12 @@ def check_id_exists_in_table(id_name: str,id_val: int, table_name: str, db: sqli
     else:
         return False
 
-# This is an example endpoint. Change the "/" to the path. For example: "/listclasses"
-# For each new endpoint here, create a new endpoint in the etc/krakend.json file
+@app.get("/table/{tablename}/{limit}") # Remove this before we turn in
+def view_table(tablename: str,limit:int, db: sqlite3.Connection = Depends(get_db)):
+    res = db.execute(f"SELECT * FROM {tablename}")
+    return {tablename: res.fetchmany(limit)}
 
-@app.get("/")
+@app.get("/") # Remove this before we turn in
 def greet():
     return {"Hello": "World"}
 
@@ -49,7 +51,7 @@ def list_open_classes(db: sqlite3.Connection = Depends(get_db)):
     )
     return {"Classes": classes.fetchall()}
 
-@app.post("/enroll/{studentid}/{classid}/{sectionid}") # Janhvi
+@app.post("/enroll/{studentid}/{classid}/{sectionid}", status_code=status.HTTP_201_CREATED) # Janhvi
 def enroll_student_in_class(studentid: int, classid: int, sectionid: int, db: sqlite3.Connection = Depends(get_db)):
     class_data = db.execute("SELECT * FROM Classes WHERE ClassID = ?", (classid,)).fetchone()
     if not class_data:
@@ -72,69 +74,87 @@ def enroll_student_in_class(studentid: int, classid: int, sectionid: int, db: sq
 
     return {"message": f"Student {studentid} enrolled in class {classid}"}
 
-@app.delete("/enrollmentdrop/{studentid}/{classid}") # Done
-def drop_student_from_class(studentid: int, classid: int, response: Response, db: sqlite3.Connection = Depends(get_db)):
+@app.delete("/enrollmentdrop/{studentid}/{classid}/{sectionid}") # Done
+def drop_student_from_class(studentid: int, classid: int,sectionid: int, response: Response, db: sqlite3.Connection = Depends(get_db)):
     # Try to Remove student from the class
-    try:
-        dropped_student = db.execute(f"SELECT StudentID FROM Enrollments WHERE StudentID = {studentid} AND ClassID = {classid}").fetchone()
-    except sqlite3.IntegrityError as e:
+    
+    dropped_student = db.execute("SELECT StudentID FROM Enrollments WHERE StudentID = ? AND ClassID = ? AND SectionNumber = ?",(studentid,classid,sectionid)).fetchone()
+    if not dropped_student:
         raise HTTPException(
-            status_code=status.HTTP_40,
-            detail={"ErrorType": type(e).__name__, "ErrorMessage": str(e)}
-        )
-   
-    query = db.execute(f"UPDATE Enrollments SET EnrollmentStatus = 'DROPPED' WHERE StudentID = {studentid} AND ClassID = {classid}")
+            status_code=status.HTTP_404_NOT_FOUND, 
+            detail="Student and class combination not found")
+
+    query = db.execute("UPDATE Enrollments SET EnrollmentStatus = 'DROPPED' WHERE StudentID = ? AND ClassID = ?", (studentid, classid))
     db.commit()
-    # Add student to class if there are students in the waitlist for this course
-    next_on_waitlist = db.execute(f"SELECT * FROM Waitlists WHERE ClassID = {classid} ORDER BY Position ASC").fetchone()
+    # Add student to class if there are students in the waitlist for this class
+    next_on_waitlist = db.execute("SELECT * FROM Waitlists WHERE ClassID = ? ORDER BY Position ASC", (classid,)).fetchone()
     if next_on_waitlist:
-        db.execute(f"DELETE FROM Waitlists WHERE StudentID = {studentid} AND ClassID = {classid}")
-        db.execute(f"  INSERT INTO Enrollments(StudentID, ClassID, SectionNumber,EnrollmentStatus) \
-                       VALUES ({studentid}, {classid}, {dropped_student['SectionNumber'],'ENROLLED'})")
-        db.execute("UPDATE Classes SET WaitlistCount = WaitlistCount + 1 WHERE ClassID = ?", (classid,))
-        db.commit()
+        try:
+            db.execute("INSERT INTO Enrollments(StudentID, ClassID, SectionNumber,EnrollmentStatus) \
+                            VALUES (?, ?, ?,'ENROLLED')", (next_on_waitlist['StudentID'], classid, sectionid))
+            db.execute("DELETE FROM Waitlists WHERE StudentID = ? AND ClassID = ?", (next_on_waitlist['StudentID'], classid))
+            db.execute("UPDATE Classes SET WaitlistCount = WaitlistCount - 1 WHERE ClassID = ?", (classid,))
+            db.commit()
+        except sqlite3.IntegrityError as e:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail={
+                    "ErrorType": type(e).__name__, 
+                    "ErrorMessage": str(e)
+                },
+            )
+        
         return {"Result": [
             {"Student dropped from class": dropped_student}, 
             {"Student added from waitlist": next_on_waitlist},
         ]}
-    if dropped_student:
-        return {"Student dropped from class": dropped_student}
-    else:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Student and class combination not found")
-        #return {"Result": "Student and class combination not found"}
+    return {"Result": [{"Student dropped from class": dropped_student} ]}
 
-@app.delete("/waitlistdrop/{studentid}/{classid}") # Logan
+@app.delete("/waitlistdrop/{studentid}/{classid}") # Logan DONE
 def remove_student_from_waitlist(studentid: int, classid: int, db: sqlite3.Connection = Depends(get_db)):
-    exists = db.execute(f"SELECT * FROM Waitlists WHERE StudentID = {studentid} AND ClassID = {classid}").fetchone()
-    if exists:
-        db.execute(f"DELETE FROM Waitlists WHERE StudentID = {studentid} AND ClassID = {classid}")
-        db.execute("UPDATE Classes SET WaitlistCount = WaitlistCount - 1 WHERE ClassID = ?", (classid,))
-        db.commit()
-        return {"Element removed": exists}
+    exists = db.execute("SELECT * FROM Waitlists WHERE StudentID = ? AND ClassID = ?", (studentid, classid)).fetchone()
+    if not exists:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail={"Error": "No such student found in the given class on the waitlist"}
+        )
+    db.execute("DELETE FROM Waitlists WHERE StudentID = ? AND ClassID = ?", (studentid, classid))
+    db.execute("UPDATE Classes SET WaitlistCount = WaitlistCount - 1 WHERE ClassID = ?", (classid,))
+    db.commit()
+    return {"Element removed": exists}
+    
+@app.get("/waitlist/{studentid}/{classid}") # Janhvi DONE
+def view_waitlist_position(studentid: int, classid: int, db: sqlite3.Connection = Depends(get_db)):
+    position = None
+    position = db.execute("SELECT Position FROM Waitlists WHERE StudentID = ? AND ClassID = ?", (studentid,classid,)).fetchone()
+    
+    if position:
+        message = f"Student {studentid} is on the waitlist for class {classid} in position"
     else:
-        return {"Error": "No such student found in the given class on the waitlist"}
-
-@app.get("/waitlist/{studentid}/{classid}") # Janhvi
-def view_waitlist_position(studentid: int, classid: int, sectionid: int, db: sqlite3.Connection = Depends(get_db)):
-    try:
-        waitlist_data = db.execute("SELECT Position FROM Waitlists WHERE StudentID = :studentid AND ClassID = :classid", {"studentid": studentid, "classid": classid}).fetchone()
-        position = waitlist_data["Position"]
-        message = f"Student {studentid} is on the waitlist for class {classid}"
-    except TypeError:
         message = f"Student {studentid} is not on the waitlist for class {classid}"
-    return {"message": message}
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=message,
+        )
+    return {message: position}
     
 ### Instructor related endpoints
 
-@app.get("/enrolled/{instructorid}/{classid}/{sectionid}") # Logan
+@app.get("/enrolled/{instructorid}/{classid}/{sectionid}") # Logan DONE
 def view_enrolled(instructorid: int, classid: int, sectionid: int, db: sqlite3.Connection = Depends(get_db)):
-    enrolled = db.execute(f"SELECT StudentID FROM Enrollments as e INNER JOIN \
-                                (SELECT ClassID FROM Classes WHERE InstructorID = {instructorid} AND ClassID = {classid}) as ic \
-                            ON e.ClassID = ic.ClassID")
-    
-    return {"All students enrolled in this instructor's classes" : enrolled.fetchall()}
+    enrolled_students = db.execute("SELECT * FROM Students INNER JOIN (SELECT StudentID FROM Enrollments as e INNER JOIN \
+                                    (SELECT ClassID FROM Classes WHERE InstructorID = ? AND ClassID = ? AND SectionNumber = ?) as ic \
+                                ON e.ClassID = ic.ClassID WHERE e.EnrollmentStatus = \"ENROLLED\") as si \
+                            ON Students.StudentID = si.StudentID",(instructorid,classid,sectionid))
+    enrolled = enrolled_students.fetchall()
+    if enrolled:
+        return {"All students enrolled in this instructor's classes" : enrolled}
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_204_NO_CONTENT,
+        )
 
-@app.get("/instructor/dropped/{instructorid}/{classid}/{sectionid}")
+@app.get("/dropped/{instructorid}/{classid}/{sectionid}") # Arjit DONE
 def view_dropped_students(instructorid: int, classid: int, sectionid: int, db: sqlite3.Connection = Depends(get_db)):
     query = "SELECT StudentID FROM Enrollments WHERE ClassID = ? AND SectionNumber = ? AND EnrollmentStatus = 'DROPPED'"
     dropped_students = db.execute(query, (classid, sectionid)).fetchall()
@@ -142,7 +162,7 @@ def view_dropped_students(instructorid: int, classid: int, sectionid: int, db: s
         raise HTTPException(status_code=404, detail="No dropped students found for this class.")
     return {"Dropped Students": [student["StudentID"] for student in dropped_students]}
 
-@app.delete("/instructor/drop/{instructorid}/{classid}/{studentid}")
+@app.delete("/drop/{instructorid}/{classid}/{studentid}") #Arjit DONE
 def drop_student_administratively(instructorid: int, classid: int, studentid: int, db: sqlite3.Connection = Depends(get_db)):
     query = "UPDATE Enrollments SET EnrollmentStatus = 'DROPPED' WHERE StudentID = ? AND ClassID = ?"
     result = db.execute(query, (studentid, classid))
@@ -151,24 +171,37 @@ def drop_student_administratively(instructorid: int, classid: int, studentid: in
         raise HTTPException(status_code=404, detail="Student, class, or section not found.")
     return {"message": f"Student {studentid} has been administratively dropped from class {classid}"}
 
-@app.get("/waitlist/{instructorid}/{classid}/{sectionid}")
+@app.get("/waitlist/{instructorid}/{classid}/{sectionid}") #Arjit DONE
 def view_waitlist(instructorid: int, classid: str, sectionid: int, db: sqlite3.Connection = Depends(get_db)):
-    query = "SELECT StudentID, Position FROM Waitlists WHERE ClassID = ? AND SectionNumber =? AND InstructorID = ? ORDER BY Position"
+    query = "SELECT * FROM Students INNER JOIN (SELECT StudentID, Position FROM Waitlists WHERE ClassID = ? AND SectionNumber =? AND InstructorID = ? ORDER BY Position) as w on Students.StudentID = w.StudentID"
     waitlist = db.execute(query, (classid, sectionid, instructorid)).fetchall()
     if not waitlist:
-        raise HTTPException(status_code=404, detail="No students found in the waitlist for this class.")
-    return {"Waitlist": [{"student_id": student["StudentID"], "position": student["Position"]} for student in waitlist]}
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No students found in the waitlist for this class")
+    return {"Waitlist": [{"student_id": student["StudentID"], "Email": student["Email"], "Name": student["Name"],"position": student["Position"]} for student in waitlist]}
 
 ### Registrar related endpoints
 
-@app.post("/add/{classid}/{sectionid}") # Melissa
-def add_class(classid: str, sectionid: str, db: sqlite3.Connection = Depends(get_db)):
-    db.execute("INSERT INTO Classes (ClassID,SectionNumber) VALUES(classid, sectionid)")
-    db.commit
+@app.post("/add/{classid}/{sectionid}/{instructorid}", status_code=status.HTTP_201_CREATED) # Melissa DONE
+def add_class(classid: str, sectionid: str, instructorid: str, db: sqlite3.Connection = Depends(get_db)):
+    try:
+        db.execute(f"INSERT INTO Classes (ClassID,SectionNumber, InstructorID) VALUES({classid}, {sectionid}, {instructorid})")
+        db.commit()
+    except sqlite3.IntegrityError as e:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail={
+                "ErrorType": type(e).__name__, 
+                "ErrorMessage": str(e)
+            },
+        )
     return {"New Class Added":f"Course {classid} Section {sectionid}"}
 
-@app.delete("/remove/{classid}/{sectionid}") # Melissa
+@app.delete("/remove/{classid}/{sectionid}") # Melissa DONE
 def remove_class(classid: str, sectionid: str, db: sqlite3.Connection = Depends(get_db)):
+
+    class_found = db.execute("SELECT * FROM Classes WHERE ClassID = ? AND SectionNumber = ?",(classid,sectionid)).fetchone()
+    if not class_found:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Class {classid} Section {sectionid} does not exist in database.")
     db.execute("DELETE FROM Classes WHERE ClassID =? AND SectionNumber =?", (classid, sectionid))
     db.commit()
     return {"Removed" : f"Course {classid} Section {sectionid}"}
@@ -184,16 +217,7 @@ def freeze_enrollment(isfrozen: str, db: sqlite3.Connection = Depends(get_db)):
     else:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Freeze must be true or false.")
 
-@app.get("/checkfrozen")
-def check_frozen_status(db: sqlite3.Connection = Depends(get_db)):
-    checkFrozen = db.execute("SELECT IsFrozen FROM Freeze").fetchone()
-    if (checkFrozen[0] == 1):
-        checkFrozen = True
-    else:
-        checkFrozen = False
-    return {"Enrollment Frozen": checkFrozen}
-
-@app.put("/change/{classid}/{newprofessorid}", status_code=status.HTTP_204_NO_CONTENT) # Done
+@app.put("/change/{classid}/{newprofessorid}", status_code=status.HTTP_204_NO_CONTENT) # Micah Done
 def change_prof(classid: int, newprofessorid: int, db: sqlite3.Connection = Depends(get_db)):
     valid_instructor_id = check_id_exists_in_table("InstructorID",newprofessorid,"Classes",db)
     valid_class_id = check_id_exists_in_table("ClassID",classid,"Classes",db)
